Fix delegate calls when data is changed
This illuminates the problem of who owns which data, in this case
EventTimerControl gets called with a starting date and after that it
owns the date until stop is called. But when stop is called a date is
not supplied, which is wrong since the EventViewModel for EventViewController
actually sets the stopdate and then notifies the EventViewController of
that, but in this case EventTimerControl created a stop date also and
notified EventViewController of that, which in turn notified the
EventViewModel. This in turn lead to more updates of the CoreData object
than what was needed. This fix makes sure that a stop date is called
into EventTimerControl and with some careful placement of when
a delegate is called, in this case only when the component changes the
date, fixes the problem of ownership of the date.